Consolidate AssignmentProgressIndividual naming
