+++ src/app/components/pages/assignment_progress/QuizTeacherFeedback.tsx
@@ -49,8 +49,8 @@ import {
 import {FetchBaseQueryError} from "@reduxjs/toolkit/query";
 import {SerializedError} from "@reduxjs/toolkit";
 import {ShowLoadingQuery} from "../../handlers/ShowLoadingQuery";
-import {markClassesInternal} from "./AssignmentProgressIndividual";
-import {AssignmentSummaryCard} from "./SingleAssignmentProgress";
+import {markClassesInternal} from "./AssignmentProgressIndividualTabs";
+import {AssignmentSummaryCard} from "./AssignmentProgressIndividual";
 import classNames from "classnames";
 import {Spacer} from "../../elements/Spacer";
 import { ResultsTableHeader } from "./ResultsTableHeader";
